create: clogger/levelMap
this is where the level map type and its functions is defined,
along with some static variables that implement the map.
